refactor: Use common types package for all runtimes
