+++ runtime/mock/Cargo.toml
@@ -21,11 +21,8 @@ serde       = { version = "1.0",  optional = true, features = [ "derive" ] }
 t3rn-primitives = {  path = "../../primitives" }
 t3rn-protocol   = { path = "../../protocol"}
 
-circuit-runtime-types = {   path = "../common-types" }
-circuit-runtime-pallets = {   path = "../common-pallets" }
-
-# branch frontier-para uses polkadot-v0.9.19
-#pallet-evm-precompile-simple = { git = "https://github.com/t3rn/frontier", branch = "polkadot-v0.9.19"}
+circuit-runtime-types = { path = "../common-types" }
+circuit-runtime-pallets = { path = "../common-pallets" }
 
 # Substrate Dependencies
 frame-executive                            = { git = "https://github.com/paritytech/substrate", branch = 'polkadot-v0.9.19'}
@@ -50,15 +47,3 @@ sp-version          = { git = "https://github.com/paritytech/substrate", branch 
 
 # Mock only
 sp-io      = {  git = "https://github.com/paritytech/substrate", branch = 'polkadot-v0.9.19' }
-
-[features]
-default = [ "std" ]
-std = [ ]
-# TODO: https://github.com/paritytech/parity-bridges-common/issues/390
-# I've left the feature flag here to test our CI configuration
-runtime-benchmarks = [
-
-]
-# Make contract callable functions marked as __unstable__ available. Do not enable
-# on live chains as those are subject to change.
-contracts-unstable-interface = [ ]
