+++ runtime/t3rn-parachain/src/parachain_config.rs
@@ -1,7 +1,6 @@
-use crate::{primitives::*, *};
+use crate::*;
 
 use smallvec::smallvec;
-
 use sp_runtime::impl_opaque_keys;
 use sp_std::prelude::*;
